requirements.txt
aliyun-log-python-sdk==0.6.22
APScheduler==3.5.1
certifi==2018.1.18
chardet==3.0.4
demjson==2.2.4
enum34==1.1.6
idna==2.6
Jinja2==2.10
loghub==0.3.0
MarkupSafe==1.0
numpy==1.14.0
pip==9.0.1
protobuf==3.5.1
py4j==0.10.4
pyspark==2.2.1
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
setuptools==38.5.1
six==1.11.0
tornado==4.5.3
tzlocal==1.5.1
urllib3==1.22
wheel==0.30.0
